requirements.txt
Django
psycopg2
djangorestframework
djangorestframework-simplejwt
Pillow
django-cors-headers
django-environ
coreapi
coreschema
uritemplate
django-filter
drf-yasg
django-rest-passwordreset
whitenoise
gunicorn
django-debug-toolbar
django-extensions

